Update MVC5-SPA-Angular/SPAuth/Web/Utility.cs
This is outside the scope of this sample, but this method can be a more generic URL handler if it could handle both relative and absolute paths.  

This way the caller does not need to worry about actual URL.  This is a very similar implementation to what I have written in the past.  In a load balanced server setup, It allows configuration to specify relative or absolute paths depending on whether services are deployed on all or individual servers in the server collection.  If the URL is relative it will scope server to server requests within the same machine.  If absolute, the requests will be routed specifically.

You can add the following logic after line 11:

if (relativeUrl.ToLower().StartsWith("http"))
    return relativeUrl;